Move `CoreDataFeedStore` implementation of `FeedImageDataStore` protocol to production.
