+++ EssentialFeed/EssentialFeedTests/Feed Cache/CoreDataFeedImageDataStoreTests.swift
@@ -8,23 +8,6 @@
 import EssentialFeed
 import XCTest
 
-extension CoreDataFeedStore: FeedImageDataStore {
-
-    public func insert(
-        _ data: Data,
-        for url: URL,
-        completion: @escaping (FeedImageDataStore.InsertionResult) -> Void
-    ) {
-    }
-
-    public func retrieve(
-        dataForURL url: URL,
-        completion: @escaping (FeedImageDataStore.RetrievalResult) -> Void
-    ) {
-        completion(.success(.none))
-    }
-}
-
 final class CoreDataFeedImageDataStoreTests: XCTestCase {
 
     func test_retrieveImageData_deliversNotFoundWhenEmpty() {
